refactor: move `KeyParser` and related utilities to `no.fintlabs.kafka` package and update imports across the codebase
